requirements.txt
itemadapter==0.9.0
python_docx==1.1.2
Scrapy==2.12.0
